refactor(artwork): move artworkItemName into core/artwork as ItemName
