+++ cmd/artwork_test.go
@@ -434,19 +434,19 @@ var _ = Describe("discArtworkName", func() {
 	})
 
 	It("names the album, the disc and its subtitle", func() {
-		name, err := artworkItemName(context.Background(), ds, model.KindDiscArtwork, "al-1:2")
+		name, err := artwork.ItemName(context.Background(), ds, model.KindDiscArtwork, "al-1:2")
 		Expect(err).ToNot(HaveOccurred())
 		Expect(name).To(Equal("Sandinista! (disc 2): Side Three"))
 	})
 
 	It("omits the subtitle when the disc has none", func() {
-		name, err := artworkItemName(context.Background(), ds, model.KindDiscArtwork, "al-1:1")
+		name, err := artwork.ItemName(context.Background(), ds, model.KindDiscArtwork, "al-1:1")
 		Expect(err).ToNot(HaveOccurred())
 		Expect(name).To(Equal("Sandinista! (disc 1)"))
 	})
 
 	It("rejects an id that is not <albumID>:<disc>", func() {
-		_, err := artworkItemName(context.Background(), ds, model.KindDiscArtwork, "al-1")
+		_, err := artwork.ItemName(context.Background(), ds, model.KindDiscArtwork, "al-1")
 		Expect(err).To(HaveOccurred())
 	})
 })
